requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
reportlab
google-generativeai
anthropic
